requirements.txt
urllib3==1.26.6
python-dotenv==1.0.0
google-cloud-bigquery==3.11.4
requests